+++ src/test/kotlin/org/kalasim/examples/bank/oneclerk/Bank1clerk.kt
@@ -2,7 +2,7 @@
 package org.kalasim.examples.bank.oneclerk
 
 import org.kalasim.*
-import org.kalasim.misc.display
+import org.kalasim.plot.kravis.display
 import org.koin.core.component.get
 import org.koin.core.component.inject