+++ src/test/kotlin/org/kalasim/examples/bank/standby/Bank3ClerksStandby.kt
@@ -1,6 +1,6 @@
 //Bank3ClerksStandby.kt
 import org.kalasim.*
-import org.kalasim.misc.display
+import org.kalasim.plot.kravis.display
 import org.koin.core.component.get
 import org.koin.core.component.inject