+++ src/test/kotlin/org/kalasim/examples/bank/threeclerks/Bank3Clerks.kt
@@ -2,7 +2,7 @@
 package org.kalasim.examples.bank.threeclerks
 
 import org.kalasim.*
-import org.kalasim.misc.display
+import org.kalasim.plot.kravis.display
 import org.koin.core.component.get
 import org.koin.core.component.inject
 import java.awt.GraphicsEnvironment.isHeadless