+++ src/test/kotlin/org/kalasim/test/DemoTests.kt
@@ -1,0 +1,48 @@
+package org.kalasim.test
+
+import jetbrains.letsPlot.geom.geom_boxplot
+import jetbrains.letsPlot.geom.geom_density
+import jetbrains.letsPlot.geom.geom_point
+import jetbrains.letsPlot.ggsize
+import jetbrains.letsPlot.intern.Plot
+import jetbrains.letsPlot.lets_plot
+import krangl.irisData
+import krangl.toMap
+import org.kalasim.demo.MM1Queue
+import org.kalasim.misc.showPlot
+import org.kalasim.plot.letsplot.display
+
+class DemoTests {
+
+}
+
+fun main() {
+    MM1Queue().apply {
+        run(100)
+        server.claimedMonitor.display().showPlot()
+        server.requesters.lengthOfStayMonitor.display().showPlot()
+    }
+}
+
+object LetsPlotASimulation {
+    @JvmStatic
+    fun main(args: Array<String>) {
+        val rand = java.util.Random()
+        val data = mapOf<String, Any>(
+            "rating" to List(200) { rand.nextGaussian() } + List(200) { rand.nextGaussian() * 1.5 + 1.5 },
+            "cond" to List(200) { "A" } + List(200) { "B" }
+        )
+
+        var p: Plot = lets_plot(data)
+        p += geom_density(color = "dark_green", alpha = .3) { x = "rating"; fill = "cond" }
+        p + ggsize(500, 250)
+//        p.show()
+        p.showPlot()
+
+
+        (lets_plot(irisData.toMap()) + geom_boxplot(){ x="Species" ; y="Petal.Length" }).showPlot()
+        (lets_plot(irisData.toMap()) + geom_point(){ x="Petal.Width" ; y="Petal.Length" }).showPlot()
+
+    }
+}
+